Creates shared getRelevantPackages
