+++ package.json
@@ -35,8 +35,8 @@
     "check:deps": "ts-node -O '{\"module\": \"commonjs\", \"resolveJsonModule\": true}' ./scripts/depcheck.ts",
     "check:ts": "tsc --project tsconfig.json",
     "clean": "yarn clean:builds && yarn clean:modules",
-    "clean:builds": "rm -rf packages/*/dist packages/*/tsconfig.tsbuildinfo",
-    "clean:modules": "rm -rf packages/*/node_modules node_modules",
+    "clean:builds": "rm -rf packages/*/dist packages/*/tsconfig.tsbuildinfo website/tsconfig.tsbuildinfo ./tsconfig.tsbuildinfo",
+    "clean:modules": "rm -rf node_modules packages/*/node_modules ",
     "demo:storybook": "yarn build:storybook && ./scripts/demo-storybook.sh",
     "docs:tsdoc": "yarn ts-node ./scripts/parse-tsdocs.ts",
     "docs:proptypes": "yarn ts-node ./scripts/proptypes.ts",
